Disallow the deletion/modification to system admin.

+++ clc/modules/www/src/main/java/com/eucalyptus/webui/server/EuareWebBackend.java
@@ -968,9 +968,13 @@ public class EuareWebBackend {
     for ( String id : ids ) {
       try { 
         User user = Accounts.lookupUserById( id );
-        Account account = user.getAccount( );
-        EuarePermission.authorizeDeleteUser( requestUser, account, user );
-        account.deleteUser( user.getName( ), false, true );
+        if ( !user.isSystemAdmin( ) ) {
+          Account account = user.getAccount( );
+          EuarePermission.authorizeDeleteUser( requestUser, account, user );
+          account.deleteUser( user.getName( ), false, true );
+        } else {
+          throw new IllegalArgumentException( "Can't delete system admin" );
+        }
       } catch ( Exception e ) {
         LOG.error( "Failed to delete user " + id, e );
         LOG.debug( e, e );
@@ -1298,6 +1302,9 @@ public class EuareWebBackend {
       }
       
       User user = Accounts.lookupUserById( userId );
+      if ( user.isSystemAdmin( ) ) {
+        throw new EucalyptusServiceException( "Can not modify system admin" );
+      }
       EuarePermission.authorizeModifyUser( requestUser, user.getAccount( ), user );
       if ( !user.getName( ).equals( userName ) ) {
         user.setName( ValueCheckerFactory.createUserAndGroupNameChecker( ).check( userName ) );
